src/test/java/com/example/jpapractice/repository/UserRepositoryTest.java
package com.example.jpapractice.repository;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
class UserRepositoryTest {

    @Autowired
    private UserRepository userRepository;

    @Test
    void save() {
        userRepository.save();
    }

    @Test
    void find() {
        String name = userRepository.findName();
        assertThat(name).isEqualTo("수현");
    }
}